64-bit / big-endian fix (sscanf requires pointer to int)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6852 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/file/file_system_util.cpp
@@ -130,9 +130,9 @@ void fs_NextNumberedFilename(const PIVFS& fs, const VfsPath& pathnameFormat, siz
 		fs->GetDirectoryEntries(path, &files, 0);
 		for(size_t i = 0; i < files.size(); i++)
 		{
-			size_t number;
+			int number;
 			if(sscanf(files[i].Name().c_str(), nameFormat.c_str(), &number) == 1)
-				maxNumber = std::max(number, maxNumber);
+				maxNumber = std::max(size_t(number), maxNumber);
 		}
 
 		nextNumber = maxNumber+1;
